moved create project dialog to deploy.ui, reload project list after dialog closes
