+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy.ui/META-INF/MANIFEST.MF
@@ -25,6 +25,7 @@ Import-Package: com.google.api.client.auth.oauth2,
  com.google.cloud.tools.eclipse.ui.util,
  com.google.cloud.tools.eclipse.ui.util.console,
  com.google.cloud.tools.eclipse.ui.util.databinding,
+ com.google.cloud.tools.eclipse.ui.util.event,
  com.google.cloud.tools.eclipse.usagetracker,
  com.google.cloud.tools.eclipse.util,
  com.google.cloud.tools.eclipse.util.status,
