+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy.ui/src/com/google/cloud/tools/eclipse/appengine/deploy/ui/messages.properties
@@ -40,6 +40,11 @@ tooltip.stop.previous.version=If checked, stops the previously running version w
  new version that receives all traffic.
 tooltip.staging.bucket=The Google Cloud Storage bucket used to stage files for deployment. \
  If not specified, the application''s default code bucket is used.
+
+#Project selector
+projectselector.create.newproject=Create GCP project...
+projectselector.dialog.createproject.title=Create new GCP project
+projectselector.dialog.createproject.message=You can create a new GCP project in the <a href="{0}">Cloud Console</a>
 projectselector.retrieveproject.error.title=Failed to retrieve projects
 projectselector.retrieveproject.error.message=An error occurred while retrieving projects: {0}
 projectselector.retrieveapplication.error.title=Failed to retrieve App Engine application
