+++ plugins/com.google.cloud.tools.eclipse.projectselector/src/com/google/cloud/tools/eclipse/projectselector/messages.properties
@@ -1,7 +1,4 @@
 projectselector.header.name=Name
 projectselector.header.id=ID
-projectselector.create.newproject=Create GCP project...
 openurllistener.error.title=Error
 openurllistener.error.message=Could not open URL
-dialog.createproject.title=Create new GCP project
-dialog.createproject.message=You can create a new GCP project in the <a href="{0}">Cloud Console</a>
